requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
aiosqlite
annotated-doc
annotated-types
anyio
attrs
banks
beautifulsoup4
cachetools
certifi
charset-normalizer
click
colorama
contourpy
cycler
dataclasses-json
deprecated
dirtyjson
distro
faiss-cpu
fastapi
filetype
fonttools
frozenlist
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
griffe
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
idna
iniconfig
jinja2
jiter
joblib
kiwisolver
llama-cloud
llama-index-agent-openai
llama-index-cli
llama-index-embeddings-gemini
llama-index-embeddings-ollama
llama-index-embeddings-openai
llama-index-indices-managed-llama-cloud
llama-index-instrumentation
llama-index-llms-gemini
llama-index-llms-ollama
llama-index-llms-openai
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-index-workflows
llama-parse
markupsafe
marshmallow
matplotlib
multidict
mypy-extensions
nest-asyncio
networkx
nltk
numpy
ollama
openai
packaging
pandas
pillow
platformdirs
pluggy
propcache
proto-plus
protobuf
pyasn1
pyasn1-modules
pydantic
pydantic-core
pygments
pyparsing
pypdf
pytest
pytest-asyncio
python-dateutil
python-dotenv
python-multipart
pytz
pyyaml
regex
requests
rsa
setuptools
six
sniffio
soupsieve
sqlalchemy
starlette
striprtf
tenacity
tiktoken
tqdm
typing-extensions
typing-inspect
typing-inspection
tzdata
uritemplate
urllib3
uvicorn
wrapt
yarl
